--- apps/website/components/organisms/settings/index.ts
@@ -1,6 +1,0 @@
-import { NavBarSettings } from "./navbar-settings";
-import AccountManagementSettings from "./account-management";
-import ConnectedAccountsSettings from "./connected-accounts";
-import NotificationsSettings from "./notifications";
-
-export { NavBarSettings, AccountManagementSettings, ConnectedAccountsSettings, NotificationsSettings };
