+++ apps/website/components/templates/settings/account-management/index.ts
@@ -1,0 +1,3 @@
+import AccountManagementSettings from "./account-management";
+
+export { AccountManagementSettings };
