+++ apps/website/components/templates/settings/connected-accounts/index.ts
@@ -1,0 +1,3 @@
+import ConnectedAccountsSettings from "./connected-accounts";
+
+export { ConnectedAccountsSettings };
