+++ apps/website/components/templates/settings/index.ts
@@ -1,3 +1,13 @@
+import { AccountManagementSettings } from './account-management';
+import { ConnectedAccountsSettings } from './connected-accounts';
+import { NotificationsSettings } from './notifications';
+import { EditProfileSettings } from './profile';
 import SettingsTemplate from './SettingsTemplate';
 
-export { SettingsTemplate };
+export {
+  SettingsTemplate,
+  AccountManagementSettings,
+  ConnectedAccountsSettings,
+  NotificationsSettings,
+  EditProfileSettings
+};
