+++ apps/website/components/templates/settings/notifications/index.ts
@@ -1,0 +1,3 @@
+import NotificationsSettings from "./notifications";
+
+export { NotificationsSettings };
