+++ apps/website/components/templates/settings/profile/index.ts
@@ -1,1 +1,1 @@
-export * from './EditProfileTemplate';
+export * from './edit-profile';
