+++ apps/website/pages/settings/account-management.tsx
@@ -2,7 +2,7 @@ import { useAuth } from '../../providers/auth';
 
 import { DashboardTemplate } from '../../components/templates/dashboard';
 import { SettingsTemplate } from '../../components/templates/settings';
-import { AccountManagementSettings } from 'apps/website/components/organisms/settings';
+import { AccountManagementSettings } from '../../components/templates/settings';
 
 export default function AccountManagementSettingsPage() {
   const { me } = useAuth();
